+++ lib/scryglass/binding_tracker.rb
@@ -3,12 +3,8 @@ module Scryglass
     attr_accessor :console_binding, :user_named_variables
 
     def initialize(console_binding:)
-      self.console_binding = console_binding
       self.user_named_variables = []
-    end
-
-    def receiver
-      console_binding.receiver
+      self.console_binding = console_binding
     end
   end
 end